% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command-package.R
\docType{package}
\name{command-package}
\alias{command}
\alias{command-package}
\title{command: Process command line arguments}
\description{
Process arguments passed at the command line,
as part of data analysis workflow.
}
\details{
\itemize{
\item \code{\link[=cmd_assign]{cmd_assign()}} Process command line arguments
\item \code{\link[=extract_shell]{extract_shell()}} Turn a \code{cmd_assign()} call into a shell command
\item \code{\link[=extract_make]{extract_make()}} Turn a \code{cmd_assign()} call into a Makefile rule
\item \code{\link[=shell_script]{shell_script()}} Create a shell script
\item \code{\link[=makefile]{makefile()}} Create a Makefile
\item \href{https://bayesiandemography.github.io/command/articles/quickstart.html}{Quick Start}
How to use \code{cmd_assign()}
\item \href{https://bayesiandemography.github.io/command/articles/workflow.html}{Modular Workflows for Data Analysis}
Safe, flexible data analysis workflows
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bayesiandemography.github.io/command/}
  \item \url{https://github.com/bayesiandemography/command}
  \item Report bugs at \url{https://github.com/bayesiandemography/command/issues}
}

}
\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

Authors:
\itemize{
  \item John Bryant \email{john@bayesiandemography.com}
}

Other contributors:
\itemize{
  \item Bayesian Demography Limited [copyright holder]
}

}
\keyword{internal}
