% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filename_opt.R
\name{get_filename_opt}
\alias{get_filename_opt}
\title{Extract options from a file name or path}
\usage{
get_filename_opt(
  filename,
  which = 1,
  sep = "-",
  fixed = TRUE,
  choices = NULL,
  convert = TRUE,
  has_ext = TRUE
)
}
\arguments{
\item{filename}{File name or path.}

\item{which}{A number. Which option to extract,
starting from the left-most.}

\item{sep}{The string used to separate the base
from the options, and to separate different options.}

\item{fixed}{If \code{TRUE} (the default), match \code{sep} exactly.
If \code{FALSE}, interpret \code{sep} as a regular expression.}

\item{choices}{Valid values for the option. If \code{NULL}
(the default), any values are allowed.}

\item{convert}{Whether to convert numeric options to numbers
(via \code{\link[utils]{type.convert}}). Defaults to \code{TRUE}.}

\item{has_ext}{Whether \code{filename} includes a file extension
(e.g. \code{.rds}). Defaults to \code{TRUE}.}
}
\value{
A character or (if \code{convert} is \code{TRUE})
possibly a numeric vector of length 1.
}
\description{
Extract options from a file name or path, based
on conventions for constructing filenames.
Options include choices for inputs, parameters, or outputs.
For instance, a regression might be run several times with
different combinations of dataset and model specification.
}
\details{
\code{filename} can include directories, and
is assumed to have the form

\code{<dir>/<base><option><sep>...<sep><option><ext>}

For instance, if a model is run separately on females and males,
then \code{filename} might be

\code{out/model-female.rds}

where \code{"-"} is being used to separate the option (\code{"female"})
from the base (\code{"model"}).

If graphs are being produced for "high" and "low" variants
of a forecast, with various choices of a numeric parameter,
a target might be

\code{project/figures/fig_forecast-high-3.pdf}
}
\examples{
\dontrun{
filename_opt("out/model-female.rds")

filename_opt("project/figures/fig_forecast-high-3.pdf")
filename_opt("project/figures/fig_forecast-high-3.pdf",
             which = 2)
## leave "3" as character
filename_opt("project/figures/fig_forecast-high-3.pdf",
             which = 2,
             convert = FALSE)

## non-default value for 'sep'
filename_opt("output.urban.2022.female.csv", sep = ".")

## include validity check
filename_opt("output-high.rda", choices = c("low", "med", "high"))
\dontrun{
filename_opt("output-mid.rda", choices = c("low", "med", "high"))
}

## filename does not include an extension
filename_opt("results/output.low.55",
           which = 2,
           sep = ".",
           has_ext = FALSE)
}
}
\seealso{
Function \code{\link{get_cmd_args}} can be used
to extract file name and paths.
}
