% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target.R
\name{get_target}
\alias{get_target}
\title{Get the 'target' for the current file}
\usage{
get_target(default = NULL)
}
\arguments{
\item{default}{A value to return if the current
file is not being run from the command line.}
}
\value{
A character vector of length 1,
giving the path to a file.
}
\description{
Get the 'target' for the current file, ie get a path
to a new file that holds results from running the
current file.
}
\details{
By default, \code{get_target} assumes that (i) the
current file is being run from the command line (typically
via Rscript) and that the name of the target was passed
as a command line argument. See the example below for
and example, and see the vignette for a lot more detail.

If the current file is not being run from the command line,
then \code{get_target} will use the \code{default} argument
instead. This is useful for developing code. The existence
of a default can hide problems, however, and it is often
a good idea to remove it once the code has matured.
}
\examples{
## create an R script inside a temporary directory
cat("saveRDS(42, file = makr::get_target())", file = "current_file.R")
system("Rscript current_file.R target_file.rds")

}
\seealso{
\code{\link{get_suffix_target}}
}
