% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{is_varname_valid}
\alias{is_varname_valid}
\title{Check that 'nm' is a Valid Name for an Object in R}
\usage{
is_varname_valid(nm)
}
\arguments{
\item{nm}{A string}
}
\value{
TRUE or FALSE

noRd
}
\description{
Check that 'nm' is a Valid Name for an Object in R
}
