% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_target.R
\name{opt_target}
\alias{opt_target}
\title{Extract options from a target}
\usage{
opt_target(
  target,
  which = 1,
  sep = "-",
  fixed = TRUE,
  choices = NULL,
  convert = TRUE,
  has_ext = TRUE
)
}
\arguments{
\item{target}{File path to the output being created.}

\item{which}{A number. Which option to extract,
starting from the left-most.}

\item{sep}{The string used to separate the base
from the options, and to separate different options.}

\item{fixed}{If \code{TRUE} (the default), match \code{sep} exactly.
If \code{FALSE}, interpret \code{sep} as a regular expression.}

\item{choices}{Valid values for the option. If \code{NULL}
(the default), any values are allowed.}

\item{convert}{Whether to convert numeric options to numbers
(via \code{\link[utils]{type.convert}}). Defaults to \code{TRUE}.}

\item{has_ext}{Whether \code{target} includes a file extension
(e.g. \code{.rds}). Defaults to \code{TRUE}.}
}
\value{
A character or (if \code{convert} is \code{TRUE})
possibly a numeric vector of length 1.
}
\description{
Infer option used in creating a target from the name of that target.
}
\details{
Options include choices for inputs, parameters, or outputs.
For instance, a regression might be run several times with
different combinations of dataset and model specification.

The file name \code{target} is assumed to have the form

\code{<dir>/<base><option><sep>...<sep><option><ext>}

For instance, if a model is run separately on females and males,
then \code{target} might be

\code{out/model-female.rds}

where \code{"-"} is being used to separate the option (\code{"female"})
from the base (\code{"model"}).

Or if graphs are being produced for "high" and "low" variants
of a forecast, with various choices of a numeric parameter,
a target might be

\code{project/figures/fig_forecast-high-3.pdf}
}
\examples{
opt_target("out/model-female.rds")

opt_target("project/figures/fig_forecast-high-3.pdf")
opt_target("project/figures/fig_forecast-high-3.pdf",
           which = 2)
## leave "3" as character
opt_target("project/figures/fig_forecast-high-3.pdf",
           which = 2,
           convert = FALSE)

## non-default value for 'sep'
opt_target("output.urban.2022.female.csv", sep = ".")

## include validity check
opt_target("output-high.rda", choices = c("low", "med", "high"))
\dontrun{
opt_target("output-mid.rda", choices = c("low", "med", "high"))
}

## target does not include an extension
opt_target("results/output.low.55",
           which = 2,
           sep = ".",
           has_ext = FALSE)
}
\seealso{
The value for \code{target} is typically obtained
using function \code{\link{target}}.
}
