% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_target.R
\name{path_target}
\alias{path_target}
\title{Get the file path to the target}
\usage{
path_target(default = NULL)
}
\arguments{
\item{default}{A value to return if the current
file is not being run from the command line.}
}
\value{
A character vector of length 1,
giving the path to a file.
}
\description{
Get the path to the file that holds output from the current file.
}
\details{
When called with no arguments, \code{path_target} assumes that (i) the
current file is being run from the command line,
and that the name of the target was passed
as a command line argument. See the examples below
and the vignette for this package.

If the current file is not being run from the command line,
and if a \code{default} argument has been supplied, then
\code{path_target} will use that instead. Defaults are
helpful when developing code, but can disguise problems,
so should be removed when the code has matured.
}
\examples{
## -- Run from the command line ------------------------

\dontrun{
## a file called 'foo.R' contains the lines
library(makr)
product <- 6 * 7
path_target <- path_target()
saveRDS(product, file = path_target)

## we have a shell script that contains the line
Rscript foo.R bar.rds

## running the shell script creates a
## file 'bar.rds' consisting of the number 42

## (see the article 'makr' for more complete examples)
}

## -- Use interactively --------------------------------

path_target(default = "product-high.rds")
}
\seealso{
\code{\link{filename_opt}}
}
