% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target}
\alias{target}
\title{Get the 'target' for the current file}
\usage{
target(default = NULL)
}
\arguments{
\item{default}{A value to return if the current
file is not being run from the command line.}
}
\value{
A character vector of length 1,
giving the path to a file.
}
\description{
Get the 'target' for the current file, ie get a path
to the file that holds the output created by the
current file.
}
\details{
When called with no arguments, \code{target} assumes that (i) the
current file is being run from the command line,
and that the name of the target was passed
as a command line argument. See the examples below
and the vignette for this package.

If the current file is not being run from the command line,
and if a \code{default} argument has been supplied, then
\code{target} will use that instead. Defaults are
helpful when developing code, but can disguise problems,
so should be removed when the code has matured.
}
\examples{
## -- Run from the command line ------------------------

\dontrun{
## a file called 'foo.R' contains the lines
library(makr)
product <- 6 * 7
target <- target()
saveRDS(product, file = target)

## we have a shell script that contains the line
Rscript foo.R bar.rds

## running the shell script creates a
## file 'bar.rds' consisting of the number 42

## (see the article 'makr' for a more complete example)
}

## -- Use interactively --------------------------------

target(default = "product-high.rds")
}
\seealso{
\code{\link{opt_target}}
}
