% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command-package.R
\docType{package}
\name{command-package}
\alias{command}
\alias{command-package}
\title{command: Process command line arguments}
\description{
Process arguments passed at the command line,
as part of data analysis workflow.
}
\details{
\itemize{
\item \code{\link[=cmd_assign]{cmd_assign()}} is the main function.
\item \href{https://bayesiandemography.github.io/command/articles/quickstart.html}{Quick Start Guide}
shows how to use \code{cmd_assign()}.
\item \href{https://bayesiandemography.github.io/command/articles/workflow.html}{A Workflow for Data Analysis}
presents an approach to the design of data analysis workflows.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bayesiandemography.github.io/command/}
}

}
\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

Other contributors:
\itemize{
  \item Bayesian Demography Limited [copyright holder]
}

}
\keyword{internal}
