% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command-package.R
\docType{package}
\name{command-package}
\alias{command}
\alias{command-package}
\title{command: Process command line arguments}
\description{
Process arguments passed at the command line,
and make scripts behave a little like functions.
}
\details{
The package contains a single function: \code{\link[=cmd_assign]{cmd_assign()}}.

The vignette \code{vignette("command", package = "command")}
gives examples, and discusses how \code{cmd_assign()} can
be used to make a data analysis workflow safer
and more modular.
}
\keyword{internal}
